crawler/requirements.txt
crawl4ai==0.8.0
aiohttp==3.13.3
aiofiles==25.1.0
python-dotenv==1.2.2
supabase==2.15.2
